requirements.txt
penman>=1.1.0
transformers==4.17.0
scipy==1.5.4
datasets==1.18.0
pandas==1.1.5
scikit-learn==0.24.0
prettytable==2.1.0
gradio
torch
#torch==1.8.1+cu101 -f https://download.pytorch.org/whl/torch_stable.html
setuptools==49.3.0
laserembeddings
faiss-gpu==1.6.1
pycountry
#tensorflow
#tensorflow-text
